refactored flowControl practice
